SListMemoryPool, TlsIndex
SListMemoryPool 메모리 단편화가 생길 가능성이 잇따.

각 쓸 인자마다 TlsIndex 를 alloc 해야함

+++ XMachinaServer/ServerLib/OverlappedObject.cpp
@@ -2,7 +2,7 @@
 #include "OverlappedObject.h"
 
 /* Overlapped Object */
-OverlappedObject::OverlappedObject(OverlappedIO::Type ioType) : mIoType(ioType)
+OverlappedObject::OverlappedObject(OverlappedIO::Type ioType) : mIoType(static_cast<UINT8>(ioType))
 {
 	::ZeroMemory(this, sizeof(OVERLAPPED));
 }
